refactor: implement dual-bitmask CLI validation and update clib
- introduce opt_mask_t with 'enable' and 'must' bitfields to distinguish between optional (Hybrid) and required (Path) Level 3 arguments
- update clib include files and update implementation
- reorg cmake for commons

+++ common/include/diagnostic/logging/log.h
@@ -5,7 +5,7 @@
 #pragma once
 
 typedef struct log_stats_s {
-    uint64_t sum;
+    uint32_t sum;
     uint32_t drop_cnt;
 } log_stats_t;
